node/agents/heiwaclaw/requirements.txt
watchdog==4.0.0
psutil==5.9.8
nats-py==2.8.0
httpx==0.27.0
discord.py==2.4.0
